[Autotest] Fix opt_in() flake source.
Poll for the web_views context instead of just getting it.  Other
implementations which use this function do that too.  This change
should stop the occasional flake we see in a test which uses this
function.

Also add various docstrings to avoid repo upload errors and adhere
to style guide.

TEST=ran a test 10x times with no failure
BUG=b:30208840

Change-Id: I33dc17e56f179d591ae886452bc7554a6d8cd87e
Reviewed-on: https://chromium-review.googlesource.com/361619
Commit-Ready: Katherine Threlkeld <kathrelkeld@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>