[autotest] Pass in afe_host when staging ssp.
When running stage_server_side_package, only a list of machine names is
passed through and not a list of machine dicts.  This causes the
afe_utils methods to fail when grabbing the attributes/labels since they
by default depend on the afe_host in the host object.  Update the
staging to grab the afe_host object when running in the lab.

BUG=chromium:629702
TEST=not yet

Change-Id: I59c7732c78215537893567376cee982b68bc1647
Reviewed-on: https://chromium-review.googlesource.com/361539
Reviewed-by: Dan Shi <dshi@google.com>
Commit-Queue: Dan Shi <dshi@google.com>
Tested-by: Dan Shi <dshi@google.com>