adb_host: Sync the UTC time for M and later devices.
The "-u" option in the "date" command sets the UTC time instead of
localtime used by the adb host. This patch synchronizes the time of
the adb host in the UTC timezone with the autotest server time in the
UTC timezone so they match. The previous code was syncing the localtime
ignoring the timezones.

BUG=b:30265445
TEST=syntax check; manually test a "date -u" command.

Change-Id: I632537501c0dca4e8693d09b11d0674dcabb8df8
Reviewed-on: https://chromium-review.googlesource.com/362071
Commit-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Bindu Mahadev <bmahadev@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Reviewed-by: Bryan Lewandowski <bryanlew@google.com>