[autotest] Ignore expected servo update error during deploy
During deploy process, the servo may not be configured with a board,
which is needed for the servo update.  This is an expected situation;
however, it will raise an exception that was not caught.

The sleep() before calling verify() on the servo is increased to be more
reliable.

BUG=chromium:626071
TEST=Reproduced bug and re-ran with fix

Change-Id: Id28cb4af62af9316ace5b8ebf51ce099dd4e5ec5
Reviewed-on: https://chromium-review.googlesource.com/364649
Commit-Ready: Allen Li <ayatane@google.com>
Tested-by: Allen Li <ayatane@google.com>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>