[autotest] Change host creation in deployment_test.
This refactoring is for the sake of upcoming changes relating
to servo repair.  Additionally, it allows for cleaning up
certain awkward design aspects in deployment_test.

The change switches creating hosts in the AFE from invoking the
'atest' command in a subprocess to making RPC calls directly.
The result is that the AFE host entry for each new DUT can be
passed when calling hosts.create_host() to create the CrosHost
object.  This eliminates the need to monkey patch private methods
in the host object.

Additionally, the change now guarantees that the 'servo_host' and
'servo_port' attributes for the host are added to the AFE database,
in a step towards eliminating the hard-coding of the mapping of DUT
hostnames to servo hostnames during ServoHost construction.

BUG=chromium:630018
TEST=Run the command against a spare DUT in the test lab.

Change-Id: Ia8accaf5de0076ddb68de1d831cd7c803a735bd5
Reviewed-on: https://chromium-review.googlesource.com/366843
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>