Adding dependency on "reading_light" for RTC Webcam test.
Webcam based WebRTC tests were failing in the lab because of the camera
receiving inadequate amount of light. We've added USB reading lights to
a small sample of devices and added the label "reading_light" to them
using 'atest'. Creating a new control file that adds a dependency to
this reading_light label so we can observe the test results for  a few
days before rolling out more broadly or until we have an alternate
strategy in place. Modifying the original control file is not a great
option because then the test will run on a very small number of devices,
which isn't what we want.

BUG=chromium:636146
TEST=Tested on chromeos2-row3-rack8-host15.

Change-Id: I43270bb4d63fa6817caa453201630620e5a1c928
Reviewed-on: https://chromium-review.googlesource.com/367440
Commit-Ready: Vinayak Suley <vsuley@chromium.org>
Tested-by: Vinayak Suley <vsuley@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>