[autotest] Make sure servo port is passed in as an int.
During ServoHost creation, one of the first things it'll do is setup the
ssh tunnel for servod (if config enables it) using the xmlrpc_connect method.
If the port is a string, it'll complain about that.  Make sure that the
servo port from host attributes is an int.

BUG=None
TEST=ran platform_ServoPowerStateController on test server with ssh
tunnel enabled successfully.

Change-Id: Ied249581488b9fd8bbb800029dff8449950c6bf4
Reviewed-on: https://chromium-review.googlesource.com/370240
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@google.com>