Remove support for run-by-value and the value argument.
Version 1 of the Enterprise Base Class allowed a caller to run a
test case based on the expected policy value. This feature added
unwanted complexity to the framework, and did not work reliably
(the string value had to match exactly what was shown on the
chrome://policy page, but was not the same as what was sent to the
fake DM Server.

Support for run-by-value, and entire value argument, is removed by
this CL in version 2. Note that there are three tests in the EE2E
test framework that still run tests BY_VALUE. These should updated
to instead be run by test case name.

BUG=chromium:637432
TEST=None

Change-Id: I8bcc6e7ea94e1f69da6af5affb7ff52844fd3ecc
Reviewed-on: https://chromium-review.googlesource.com/371563
Tested-by: Scott Cunningham <scunningham@chromium.org>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: Scott Cunningham <scunningham@chromium.org>