[autotest] Update exceptions typically raised by verify().
When verify() fails, typically it raises AutoservVerifyDependencyError.
Previously, such exceptions would contain the description of every
verifier that failed; this was useful for the debug logs.  Upcoming
changes for servo repair will sometimes need to print the exceptions
raised by verify(); for the exceptions raised by verify() to be
useful for printing error messages, they need to contain the actual
errors.

This changes verify and repair so that their exceptions contain both
the description and the exception message.  If the exception is
converted to a string, (i.e. for printing) it includes all the error
messages.  Logging continues to use the verifier description.

This also tweaks the logging messages to be slightly more readable.

BUG=chromium:630018
TEST=Unit tests, plus tested the new servo repair with this code

Change-Id: I3988e9cd7422e7be5d9280b93589ca63184a450e
Reviewed-on: https://chromium-review.googlesource.com/371565
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>