power_LoadTest: Change 1hour acok to wired.
Can not necessarily guarantee wificell access in the lab and with
ethernet connected the test is always failing w/ 'Ethernet interface is
active. Please remove Ethernet cable.'

Changing to ignore network as motivation for this test is to start
regularly getting c-state utilization in the absence of battery
discharge results.

BUG=chromium:616153
TEST=test_that power_LoadTest.WIRED_1hr_acok

Change-Id: I5f5da3f3fdf54ecc203ee3b076ce351cb0c4c920
Reviewed-on: https://chromium-review.googlesource.com/372198
Commit-Ready: Todd Broch <tbroch@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>