[Autotest]Adding autotest for Android Packet Filtering
The test set's up a router in 802.11g mode and connectes the phone
to it. The test assumes that the AP has scapy installed and uses it
to generate and send duplicate and unique RAs to the phone. It then
verifies how the phone behaves under differnt situations when the
RA is sent to it.

BUG=None
TEST=Tested locally

Change-Id: I12099d6dd9263786845301fd94f029a20f9796e9
Reviewed-on: https://chromium-review.googlesource.com/372378
Commit-Ready: Bindu Mahadev <bmahadev@chromium.org>
Tested-by: Bindu Mahadev <bmahadev@chromium.org>
Reviewed-by: Kris Rambish <krisr@chromium.org>