Replace .width with .naturalWidth in image-blocked test.
Use the .naturalWidth of the image, not the .width (or .clientWidth),
to determine whether Chrome is displaying the image on the page.

Currently, policy_ImagesAllowedForUrls & policy_ImagesBlockedForUrls
tests use the following to determine whether Chrome is displaying
the image on the page:

    image_is_blocked = tab.EvaluateJavaScript(
            "document.getElementById('kittens_id').width")
            < self.MINIMUM_IMAGE_WIDTH

It turns out this is not 100% reliable. Sometimes it works, but other
times, a blocked image still takes up an equivalent amount of space,
and this test fails.

The fix is to use the naturalWidth property, which is 0 when the
image is blocked:

    image_is_blocked = tab.EvaluateJavaScript(
            "document.getElementById('kittens_id').naturalWidth")
            < self.MINIMUM_IMAGE_WIDTH

BUG=chromium:639047
TEST=Run test_that <IP> policy_ImagesAllowedForUrls and
policy_ImagesBlockedForUrls.

Change-Id: I9e7410b12300a02a8f1b5ecf0cd4ef8d4e36434b
Reviewed-on: https://chromium-review.googlesource.com/373159
Commit-Ready: Scott Cunningham <scunningham@chromium.org>
Tested-by: Scott Cunningham <scunningham@chromium.org>
Reviewed-by: Scott Cunningham <scunningham@chromium.org>
Reviewed-by: danny chan <dchan@chromium.org>