Attempt to fix CFM session stress flakiness
Reducing the number of iterations to 100 and changing the timeout values
for the length of the hangout session as well as intervals between
different sessions. This might improve the flakiness but we are not
fully sure, so want to give it a try and see how it goes.

BUG=None
TEST=None

Change-Id: Ie8ed4c38562ca7338053b114b4fe7dbeacde7fdd
Reviewed-on: https://chromium-review.googlesource.com/373084
Commit-Ready: Harpreet Grewal <harpreet@chromium.org>
Tested-by: Harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>