[autotest] For ServoHost, skip servo_board when there's no AFE.
In server.hosts.servo_host.create_servo_host(), we try to extract a
board for use when constructing the servo host.  We do this by
querying the AFE.  However, there are cases where a servo can be
found, but the board can't (notably, when running `test_that`
against a DUT in the lab).

This changes create_servo_host() not to fail if the board comes
back as `None`.

BUG=chromium:640405
TEST=Run test_that with a servo test against a DUT in the lab.

Change-Id: I3cf6c525e3e136efc82f2c7a0c2cd06a38683983
Reviewed-on: https://chromium-review.googlesource.com/374502
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>