provision_AutoUpdate.double needs 30 minutes to run
The suite limit of 20 minutes was not quite enough time for the test to
run in consistently. The sheriffs/gardeners were seeing occasional red
builds because this test was aborted due to max_runtime being reached.

Typically the logs showed the test was in the process of wrapping up.
About half the time the test was being aborted during post-test cleanup
after the test itself had otherwise completed successfully.

There was no way to set the max_runtime for the test itself, but I left
a quick comment about what to look for, and what needs to be adjusted.

This patch extends the allowed time for all test in the bvt-inline and
push_to_prod suites by ten minutes (to 30 minutes), which is enough
time to get this test to pass with some room to spare.

BUG=chromium:593423
TEST=Trybot

Change-Id: I616e5619e423325f6dfbd8431d4e5f663678e6ef
Reviewed-on: https://chromium-review.googlesource.com/375201
Tested-by: Lloyd Pique <lpique@google.com>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>