[autotest] Update deployment to support duts with servo v4.
Deployments that will deploy with servo v4 will have additional data
required which will be stored in a csv file.  I've updated the
deployment script to take in a csv file and create the host in the AFE
with the appropriate details in the host attributes.

CQ-DEPEND=CL:388674
BUG=chromium:599533
TEST=locally on moblab, add hosts successfully.

Change-Id: Ib56aa6c1108d6a24bf3f7037044c5388cdca6266
Reviewed-on: https://chromium-review.googlesource.com/376841
Commit-Ready: Kevin Cheng <kevcheng@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>