BasePackageManager: Use `test` instead of `ls`.
Prior to this change, it could be confusing when there was an error or
not when fetching packages.  The fetch_pkg method runs a `ls` on the
package file to see if it already exists.  When the package doesn't
exist, or is out of date, the new package is fetched from a repo.

`ls` would return an error saying "no such file or directory" when the
package doesn't exist.  However, this is to be expected in many cases.

This commit changes the method in checking for the existence of files
and/or directories to using `test` instead of `ls`.  This provides much
quieter output.

Additionally, the log level for successfully retrieving the package was
raised to INFO.

BUG=chromium:639005
BRANCH=None
TEST=run tryjob with HWTest and verify that `ls` error messages are
no longer present.

Change-Id: I59ace7449294fadd44f1b35e1cdbba060510cd8e
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/376999
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>