abstract_ssh.py: allow overall timeout and ssh connect timeout to differ
This makes wait_up() try ssh connections multiple times before
it gives up.  The previous logic made the overall timeout identical
to the ssh connect timeout, so if the connection hung for some
reason, ssh would not be retried.  We see a situation (see BUG)
in which an ssh connection hangs for 3 minutes, while a previous
connection worked fine.  This is a workaround for that problem.
(We checked networking logs and there is no change in network
connectivity, which might explain this situation, so we don't
really know what goes on).

BUG=chromium:641424
TEST=will run test_that

Change-Id: If19dcc2702a837f0620b8c097ed82142571305a3
Reviewed-on: https://chromium-review.googlesource.com/379537
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>