[autotest] Bug fix in container hostname
CL 376221 introduced a bug in server-side packaging running in Moblab.
In Moblab, the hostname of duts are IP address, and the Moblab functions as a
dhcp server for the duts. For the container to have the same name as the dut
it's testing will confuse the dhcp server.
Also, the container's hostname must start with `test_` to allow DHCP server in
Moblab to set a correct lease time (20min)

This change modify the container name to be a string different from the dut's
hostname. Also, replace dot in the hostname with understore to avoid dns
problem.

BUG=chromium:642783
TEST=in a moblab, sudo python site_utils/lxc_functional_test.py

Change-Id: I9378b4a23f079f54f130e779a723f33c79635dab
Reviewed-on: https://chromium-review.googlesource.com/379613
Commit-Ready: Dan Shi <dshi@google.com>
Tested-by: Dan Shi <dshi@google.com>
Reviewed-by: Dan Shi <dshi@google.com>