autotest: Fix ArcTest cleanup() failures if setup() is not run
ArcTest misuses setup(), which is not guaranteed to be run on DUT. For
example if emerge builds binaries in src/ on host machine, setup() would
be skipped on DUT.
ArcTest needs a pure client side setup step when run_once() is called to
do things like Android packages installation. This CL makes arc_setup()
such a function instead of setup(). That is,
1. arc_setup() is called by default in default implementation of
initialize() of ArcTest. It tries to install needed packages.
2. Subclasses can override arc_setup() to install their specific
packages or whatever needs to be done after Android is booted.

BUG=chromium:623790
TEST=Run all the modified tests on my minnie. All passed except some tests
which fail even on a clean checkout.
CQ-DEPEND=CL:*284358

Change-Id: I16d9ccd07ab36b875d273d6915dee89318043801
Reviewed-on: https://chromium-review.googlesource.com/383454
Commit-Ready: Cheng-Yu Lee <cylee@chromium.org>
Tested-by: Cheng-Yu Lee <cylee@chromium.org>
Reviewed-by: Chung-yih Wang <cywang@google.com>