test_push: replace the unobvious powerwash with obvious powerwash test
In order to test the repairt workflow, there is a powerwash scheduled on
shards before running any tests. However, this powerwash is not a test,
so there is no clues on the afe to tell there is a powerwash on the
host. Replace it with a powerwash test, so we have logs for it.

BUG=chromium:619107
TEST=Test on testing push master.

Change-Id: I82d238fb992375ab8c9afab2e35ed2bc1853d08d
Reviewed-on: https://chromium-review.googlesource.com/384614
Commit-Ready: Shuqian Zhao <shuqianz@chromium.org>
Tested-by: Shuqian Zhao <shuqianz@chromium.org>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>