[autotest]: always run the update commands when deploying server
Now, the deployment logic is to run certain update commands only when
certain repos or script change. However, if the update fails in the
middle, rerun will skip certain updates since the corresponding repo is
not changed between the previous fail one and current one. Obsolete the
checking updates logic.

BUG=chromium:648425
TEST=Test on test server

Change-Id: I9a79886cdb9bcea6235e84b7d0754ef27eef10dc
Reviewed-on: https://chromium-review.googlesource.com/387207
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>