[Autotest] Delete touch gesture files (switch fully to remote).
Phase 1 was to enabling download of files from google storage.
That was completed and has been running with no problems.
For Phase 2, clean up the old method by deleting files and
removing the code that looks for a local gestures folder.

No functionality will be lost, as the files currently in google
storage are a superset of the files being deleted here.

TEST=ran on several devices and verified remote download
BUG=chromium:537473

Change-Id: Iced17230f8e3c85178716d43c251fa9362e3f358
Reviewed-on: https://chromium-review.googlesource.com/388901
Commit-Ready: Katherine Threlkeld <kathrelkeld@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>