Rewrite video_WebRtcCamera to report perf instead of hard-coded limits.
It hasn't worked very well to hard-assert that there are less than
10 black frames per run. All CrOS devices are different, and it's
much easier to deal with the differences (and act on any regressions)
if we report to perf instead.

Also clean up the error flow in the method that prints the perf
values, and rename some variables.

BUG=chromium:628630
TEST=Ran test on Link platform.

Change-Id: I6434a0d6d598d2c0cd88a034c396b0e9e3484456
Reviewed-on: https://chromium-review.googlesource.com/390812
Commit-Ready: Patrik Höglund <phoglund@chromium.org>
Tested-by: Edward Lesmes <ehmaldonado@chromium.org>
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>