[autotest] record more details in reboot statistics
The metrics are recording very large failure rates on a few platforms,
but I'm not finding the same amount of failures on the builders
(e.g. might the successive tentative repairs on some broken machines).

So record the failure details (the caller to check whether it is
test-related and the type of error) in a second metric and
add the machine name in the reboot_count metric to isolate bad DUTs.

BUG=chromium:631523 chromium:642372
TEST=manually, run locally host.reboot() with traces.

Change-Id: Ie436b822363ad760f5a23d1b1a0b65224b611b07
Reviewed-on: https://chromium-review.googlesource.com/394549
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: David Riley <davidriley@chromium.org>