security_SandboxedServices: rework status file parsing again
In order to support newer fields (like CapAmb), rework the parsing logic
to handle missing ones.  This requires passing back keyed values from the
device, and then parsing them as a dict in the test.

The advantage here is that we no longer rely on the order.

BUG=chromium:652969, chromium:654531
TEST=security_SandboxedServices still passes

Change-Id: If29d94fb6a9aaa63ba6ee4beff4165cc910cf86f
Reviewed-on: https://chromium-review.googlesource.com/400038
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>