autotest: Ensure a chaps key is generated and TPM is owned.
platform_MigrateChapsTokenClient: We need to ensure that the
TPM is owned before mounting a new cryptohome and in the
control file we need to make sure that a chaps key has
been generated before we can check if it is still present.

In pksc11.py: Fixed generate_user_key so that it outputs
'Sign: CKR_OK' when successful.

BUG=chromium:646626
TEST=Ran the autotest.

Change-Id: I792f5b4ad506e747632b664192a5e782eb884d25
Reviewed-on: https://chromium-review.googlesource.com/400640
Commit-Ready: Alexis Savery <asavery@chromium.org>
Tested-by: Alexis Savery <asavery@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>