firmware_SoftwareSync: Clear 0x200 prior to start.
In order to test software sync, software sync should actually be enabled
for the DUT.  By default, the firmware tests won't check to see if
software sync is disabled or not.  This commit will cause the
firmware_SoftwareSync test to clear the GBB flag that disables software
sync.

BUG=chrome-os-partner:57427
BRANCH=None
TEST=Set gbb flags to 0x340 for cyan.  cros_workon start autotest
--board cyan.  Run firmware_SoftwareSync against the DUT.  Verify that
the GBB flags are changed to 0x140 before the tests begins and that the
test passes.

Change-Id: I720031f4a9e50b68bdfbac6ffe3cee7ad2ac0925
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/401006
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: danny chan <dchan@chromium.org>