[autotest] Fix message for a test on unsupported hardware.
When a test with a hardware dependency was determined to be
unsupported on a given board, the message said "no DUTs with
expected labels".  That message was nominally accurate, but the
target audience (Sheriffs) didn't know what it meant.  This changes
the message to say simply "test not supported on this board".

BUG=chromium:658473
TEST=None

Change-Id: I1a9c02b20bb63b193f00b9d81d9db684f36c9887
Reviewed-on: https://chromium-review.googlesource.com/402061
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>