[autotest] Update servo host reboot to talk to cautotest directly.
I've seen the create_job afe call fail which causes the test job to fail.
We don't want that to happen so let's catch that exception and log it
and just have another dut schedule the reboot for us.

The afe create_job call fails because it looks like the 'create_job' rpc
returns with an ID that is not yet available when get_jobs is called and
so it returns an empty list and we try to index that and raise an
IndexError.  This looks to be caused by calling the shard instead of
cautotest directly so also change the afe to call cautotest as well.

https://pantheon.corp.google.com/storage/browser/chromeos-autotest-results/82555879-kevcheng/chromeos6-row1-rack2-host5/debug

BUG=chromium:599533
BUG=chromium:659720
TEST=None.

Change-Id: I77c7b2b96ffa3b7a9ea2ee7a1c0960c4f9065ba4
Reviewed-on: https://chromium-review.googlesource.com/403290
Commit-Ready: Kevin Cheng <kevcheng@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>