cheets_CTS/GTS: unify error/status reporting.
A slight change in result reporting philosophy due to wmatrix coloring
nd sponge status values. A test should really only report Pass or Fail.
We reserve Error for infra issues (outside of our code). For now we
keep TestWarn for retries, but in the future hopefully this can be also
be changed to a Pass (with explanatory text).

In the short run this change should make actionable failures more
visible on wmatrix.

BUG=b:31813384
TEST=test_that $DUT cheets_CTS.6.0_r9.arm.android.core.tests.libcore.package.harmony_java_math

Change-Id: I305c99965f415466f74ef1f9d3de3c075843318f
Reviewed-on: https://chromium-review.googlesource.com/403425
Commit-Ready: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>