[autotest] Increase MySQL connection timeout
Increasing the connection timeout should make initial connection errors
less likely.

The default value seems to be 10 seconds (in MySQL documentation), but I
think this depends on how the underlying MySQL C library was compiled.
The logs in the bug suggest a timeout of ~5 seconds.

Also, increase the overall timeout to 4 minutes to compensate for the
increased per-try timeout.  One concern is that if this is called in an
RPC, the RPC response will timeout before this connection attempt.
However, the existing 2 minute timeout already rules out this case.

BUG=chromium:660520
TEST=None

Change-Id: Ie0c33a7727408a41516764215d66f7dbee20492d
Reviewed-on: https://chromium-review.googlesource.com/405829
Commit-Ready: Allen Li <ayatane@chromium.org>
Tested-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>