[autotest] Fix ServoHost.repair() to support `silent`.
A recent change added a new `silent` parameter to repair() and
verify() to support suppressing logging in status.log.  Changes
were made to use the parameter in the servo_host module, but the
parameter wasn't fully plumbed end-to-end, leading to exceptions
when the new code was invoked.

This change adds the missing plumbing.

BUG=chromium:661201
TEST=test_that with a servo test

Change-Id: I776eced64eb9502f9249bd73009c11858d7a6da7
Reviewed-on: https://chromium-review.googlesource.com/405696
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>