Get WebRTC tests off bvt-perbuild and its huge device pool.
These tests now only really report reliable results if they run
on devices with adequate lighting. We have that in av_webcam but
not in the device pool of bvt-perbuild. Also, we can make the
WebRTC sheriff rotation monitoring rules simpler and reduce
the maintenance (since there's no way we can monitor all 30+
devices in bvt-perbuild; the 10 in av_webcam is more feasible).

This loses a minimal amount of correctness test coverage for
the devices that are in bvt_perbuild but not av_webcam; these
tests would have caught cases where WebRTC crashes or fails
before we get a call up. I think this will be addressed by
628633 though, where I'll implement some test that doesn't
look to closely at the actual camera frames. That test can
run on bvt-perbuild.

BUG=chromium:628630
TEST=Can't test; control file change

Change-Id: I30cfca6894367601321889dae6049a9fd465da0e
Reviewed-on: https://chromium-review.googlesource.com/406827
Commit-Ready: Patrik Höglund <phoglund@chromium.org>
Tested-by: Rohit Makasana <rohitbm@chromium.org>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Reviewed-by: Vinayak Suley <vsuley@chromium.org>