[autotest] deployment: Update host attributes if afe_host exists.
If we deploy a host that already exists in the AFE then we can never
update the host attributes.  This sucks because if you deploy with some
bad info (wrong servo serial, etc), you'll have to manually fix it in
the AFE.  Rerunning the deploy with updated host attributes in the csv
file won't fix the fields.  This fix allows for that to happen.

BUG=chromium:599533
TEST=Deployed to a dut that already existed on AFE and saw host
attributes updated to new values.

Change-Id: If78ec65fb8174197736358754cb43886cf0d3647
Reviewed-on: https://chromium-review.googlesource.com/406451
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>