[autotest] multimedia: add method to set system gain
Add method to set system gain.
We do not use chrome.audio API to set node gain because that gain of
node is not passed to Cras correctly.
We need to set this because on board with software gain, the default
software gain value will be too large.

BUG=chromium:661709
TEST=used with next CL on chell.

Change-Id: I81dd722782efcea975a82776db765f594a290eda
Reviewed-on: https://chromium-review.googlesource.com/406820
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>