Autotest: Let status_log_entry not raise exception in parsing error msg.
Status_log_entry will raise exception when getting an error msg which includes
bad characters, which leads to a traceback exicting in status.log. This may
prevent tko parser to parse the status log and upload it to tko database.

This CL replaces the bad characters with space, and won't raise exception for
this case.

BUG=chromium:662108
TEST=Ran python commands locally, Ran unittest.

Change-Id: I4fbb4bcc2fb8eea566e80c5fe28f229e092242bb
Reviewed-on: https://chromium-review.googlesource.com/406458
Commit-Ready: Xixuan Wu <xixuan@chromium.org>
Tested-by: Xixuan Wu <xixuan@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>