retry.py: provide more information in timeout exception message
This changes the message "call is timed out" to provide more
information, such as the name of the function that is timed
out.  In cases in which the function name is not useful (e.g. "wrapper"),
a label may be passed.

Also changes the exception message for some of the other
exceptions which used to have the "call is timed out" message.

BUG=chromium:645259
TEST=ran unit test

Change-Id: I688ea72a5e0abb5188d801f753e65bb330bdc43e
Reviewed-on: https://chromium-review.googlesource.com/407982
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>