security_SandboxedServices: make dump format compatibile with load
When developing service baselines for new products, it is useful to have
dump_services produce output that can be consumed by load_baseline. This
change adds a header and some missing fields to the output so that
the it can be used directly as a baseline.

BUG=b:32560160
TEST=test_that security_SandboxedServices, using dump output as baseline.

Change-Id: Ifd5a86eeda5e7c5cef2e9c80df08bff3b7b40575
Reviewed-on: https://chromium-review.googlesource.com/409616
Commit-Ready: Laurence Goodby <lgoodby@chromium.org>
Tested-by: Laurence Goodby <lgoodby@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>