arc: Fix arc tests by setting arc.enabled.
There is expected that arc.enabled is set while Arc is running.
There are several DCHECKs for this case which are not available
for integration tests. This is first part of fix. There is an
additional fix in ArcAuthService that starts Arc for tests
triggered by arc.enabled.

BUG=brillo:32660818
TEST=test_that $IP cheets_CTS.6.0_r12.arm.android.text.
     test_that $IP cheets_ContainerSmokeTest
     test_that $IP suite:bvt-inline
     This test passes in 2 modes. First is  without any additional
     CLs. Second with following CLs on board:
     http://crrev.com/2486393003
     http://crrev.com/2468973006

Change-Id: I1958512fb98314a7bf663d9fc4e7ec36e10da302
Reviewed-on: https://chromium-review.googlesource.com/410056
Commit-Ready: Yury Khmel <khmel@chromium.org>
Tested-by: Yury Khmel <khmel@chromium.org>
Reviewed-by: Luis Hector Chavez <lhchavez@google.com>