component_UpdateFlash: add End to End autotest.
This server test verifies that Adobe Pepper Flash is included
and can be started from system location. We also verify that
updates can be downloaded on demand from Omaha as component
update and that the update runs simple ActionScript. Finally
the update is removed to keep DUT without component bits
for tests that follow.

One open scenario is to component update a component update.
This is left as a TODO(ihf).

BUG=chromium:637346
TEST=Ran on samus.

Change-Id: If4ca3ebbaedc105137d560abb257ff76c12c5957
Reviewed-on: https://chromium-review.googlesource.com/413012
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>