Break Telephony tests in suites by carrier.
Due to suite scheduler de-duping a suite cannot be run
on multiple of the same target.  What the creates is a
suite that runs on multiple targets:
  angler, bullhead, shamu, and sailfish on att
  angler, bullhead, shamu, and sailfish on spt
  angler, bullhead, shamu, and sailfish on tmo
  angler, bullhead, shamu, and sailfish on vzw

Previously it was trying to run a single suite against everything
above which resulted in only one per target (phones) being
scheduled.

BUG=b:32877993
TEST=None

Change-Id: I60dd1b105cb725a46daa9b9bae66b503607e4417
Reviewed-on: https://chromium-review.googlesource.com/411383
Commit-Ready: Kris Rambish <krisr@chromium.org>
Tested-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Betty Zhou <bettyzhou@google.com>
Reviewed-by: Dan Shi <dshi@google.com>