Adding Left/Right internal speaker test.
Adding a new test that depends on audio boxes to test whether
the left and right internal speakers are working or not. It
can't tell whether the wrong speaker makes a sound for a given
channel, but it will be able to tell if a certain channel is
broken completely.

The test has two variations, one that plays the file using the
native audio player available through InternalSpeaker audio
source API, and another that plays wav files through the
browser. The idea is that if there is a failure on one and
not the other, figuring out which part of the audio stack to
investigate could be easier.

BUG=None
TEST=Tested on chromeos9-audiobox1-host1

Change-Id: I7cfa611441dd755bf362ab8562737047becfb6c2
Reviewed-on: https://chromium-review.googlesource.com/411374
Commit-Ready: Vinayak Suley <vsuley@chromium.org>
Tested-by: Vinayak Suley <vsuley@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>