autotest: retry staging when devserver returns invalid response.
There's a case that when devserver is overloaded and cannot handle some
requests like 'stage', it will return invalid responses, which won't be raised
as an HTTP error and leads to stage process terminated.

This CL retry such cases. When invalid response is detected, stage will be
retried after 30 seconds to 1) not add more burden on that devserver, 2) make
stage process continue, not be terminated.

BUG=chromium:615650
TEST=Locally run ds.trigger_download() with devserver, test whether it's
retried when the returned response of devserver is invalid.

Change-Id: Id6c4adf1fc055bfeace26cb236a0b23c833e19cd
Reviewed-on: https://chromium-review.googlesource.com/412150
Commit-Ready: Xixuan Wu <xixuan@chromium.org>
Tested-by: Xixuan Wu <xixuan@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>