servo: Use new connections for each servod RPC.
It seems that httplib is not threadsafe.  For certain FAFT tests such as
firmware_ECWakeSource, certain test cases would fail fairly reliably.
It turns out that there was some contention of the HTTPConnection to
servod between some of the wake methods and the mode_aware_reboot()
common method.  The wake methods have a "delay" decorator that delays
for a certain time period and spawns a new thread to issue RPCs to
servod.  This would cause "CannotSendRequest" exceptions thrown by the
httplib because a request was already in place.

This commit changes the behaviour to have each RPC to servod acquire a
new ServerProxy connection.  This way, the threads can't fight for the
same HTTPConnection.

BUG=chrome-os-partner:59753
BRANCH=None
TEST=run firmware_ECWakeSource against cave 4 times, and verify it
passes each time.

Change-Id: I904b6e6c24acbcec263fdf211e9ed2be24dd40f3
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/411992
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>